Move content entity to an entities folder
