+++ api/src/favorites/favorites.module.ts
@@ -1,6 +1,6 @@
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { Module } from '@nestjs/common';
-import { Content } from '../contents/content.entity';
+import { Content } from '../contents/entities/content.entity';
 import { CaslAbilityFactory } from '../shared/modules/casl/casl-ability.factory';
 import { Favorite } from './favorite.entity';
 import { FavoritesController } from './favorites.controller';
