+++ api/src/reports/report.entity.ts
@@ -6,7 +6,7 @@ import {
   PrimaryKey,
   Property,
 } from '@mikro-orm/core';
-import { Content } from '../contents/content.entity';
+import { Content } from '../contents/entities/content.entity';
 import { BaseEntity } from '../shared/lib/entities/base.entity';
 import { User } from '../users/user.entity';
