+++ api/src/reports/reports.module.ts
@@ -1,7 +1,7 @@
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 import type { OnModuleInit } from '@nestjs/common';
 import { Module } from '@nestjs/common';
-import { Content } from '../contents/content.entity';
+import { Content } from '../contents/entities/content.entity';
 import { CaslAbilityFactory } from '../shared/modules/casl/casl-ability.factory';
 import { User } from '../users/user.entity';
 import { ReportSearchService } from './report-search.service';
