+++ api/src/threads/thread.entity.ts
@@ -8,7 +8,7 @@ import {
   Property,
 } from '@mikro-orm/core';
 import { Transform } from 'class-transformer';
-import type { Content } from '../contents/content.entity';
+import type { Content } from '../contents/entities/content.entity';
 import { TransformCollection } from '../shared/lib/decorators/transform-collection.decorator';
 import { ContentMaster } from '../shared/lib/entities/content-master.entity';
 import { ContentMasterType } from '../shared/lib/types/content-master-type.enum';
