+++ api/src/threads/threads.module.ts
@@ -1,8 +1,8 @@
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 import type { OnModuleInit } from '@nestjs/common';
 import { Module } from '@nestjs/common';
-import { Content } from '../contents/content.entity';
 import { ContentsModule } from '../contents/contents.module';
+import { Content } from '../contents/entities/content.entity';
 import { Favorite } from '../favorites/favorite.entity';
 import { Reaction } from '../reactions/reaction.entity';
 import { Report } from '../reports/report.entity';
