+++ api/src/votes/vote.entity.ts
@@ -5,7 +5,7 @@ import {
   ManyToOne,
   PrimaryKey,
 } from '@mikro-orm/core';
-import { Content } from '../contents/content.entity';
+import { Content } from '../contents/entities/content.entity';
 import { BaseEntity } from '../shared/lib/entities/base.entity';
 import { User } from '../users/user.entity';
