+++ api/src/votes/votes.module.ts
@@ -1,6 +1,6 @@
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { Module } from '@nestjs/common';
-import { Content } from '../contents/content.entity';
+import { Content } from '../contents/entities/content.entity';
 import { CaslAbilityFactory } from '../shared/modules/casl/casl-ability.factory';
 import { Vote } from './vote.entity';
 import { VotesController } from './votes.controller';
